refactor: 유틸리티 클래스들을 LoggingFilter 내부로 통합
- MdcUtils, MdcCloseable을 LoggingFilter의 inner class로 이동
- 작은 유틸리티 클래스들의 분산을 방지하고 응집도 향상
- LoggingFilter.MdcUtils.KEY_REQUEST_ID로 AOP에서 접근 가능
- 파일 수 감소: 8개 → 6개 클래스

장점:
- 관련 기능들의 높은 응집도
- 불필요한 파일 분산 제거
- 로깅 관련 코드의 중앙화
