--- src/main/java/park/bumsiku/log/MdcCloseable.java
@@ -1,14 +1,0 @@
-package park.bumsiku.log;
-
-
-public class MdcCloseable implements AutoCloseable {
-
-    public static MdcCloseable create() {
-        return new MdcCloseable();
-    }
-
-    @Override
-    public void close() {
-        MdcUtils.clear();
-    }
-}
